pyproject.toml
[build-system]
requires = ["setuptools >= 40.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyAgxArm"
version = "1.0.0"
description = "A SDK to control Agilex arms"
readme = { file = "README.MD", content-type = "text/markdown" }
authors = [
    { name = "Rosen Yin", email = "yinruocheng321@gmail.com" }
]
license = { text = "MIT License" }
requires-python = ">=3.6"
dependencies = [
    "python-can>=3.3.4",
    "typing-extensions>=3.7.4.3",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
    "License :: OSI Approved :: MIT License"
]

[project.urls]
Homepage = "https://github.com/agilexrobotics/pyAgxArm"
Repository = "https://github.com/agilexrobotics/pyAgxArm"
ChangeLog = "https://github.com/agilexrobotics/pyAgxArm/blob/master/CHANGELOG.MD"

[tool.setuptools.packages.find]
include = [
    "pyAgxArm",
    "pyAgxArm.*",
]

[tool.setuptools.package-data]
"pyAgxArm" = ["py.typed", "*"]

[tool.setuptools]
include-package-data = true

